% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools.r
\name{addDbRows}
\alias{addDbRows}
\title{Add Rows to Database Table in Output SQLite Database (addDbRows)}
\usage{
addDbRows(db, dbOut, tableName, addERU, numToRead, numRows)
}
\arguments{
\item{db}{Directory path to input SQLite database.}

\item{dbOut}{Directory path to output SQLite database.}

\item{tableName}{Name of database table being sent from db to dbOut.}

\item{addERU}{Logical; if TRUE, adds ERU to GROUPS column of FVS tables. See dbCompile for details.}

\item{numToRead}{Number of rows to read in from database table at a time.}

\item{numRows}{Number of rows in argument tableName.}
}
\value{
None
}
\description{
Incrementally sends portions of a database table (as a data frame) to an output SQLite database.
}
\examples{
addDbRows("input.db", "output.db", "FVS_STANDINIT", TRUE, 1000, 5000)

}
