% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools.r
\name{addDbTable}
\alias{addDbTable}
\title{Add Database Table to Output SQLite Database (addDbTable)}
\usage{
addDbTable(db, dbOut, tableName, addERU)
}
\arguments{
\item{db}{Directory path and file name to input database.}

\item{dbOut}{Directory path and file name to output SQLite database.}

\item{tableName}{Name of database table in db being sent to dbOut.}

\item{addERU}{Logical; if TRUE, adds ERU to GROUPS column of FVS tables. See dbCompile for details.}
}
\value{
None
}
\description{
Sends an entire database table (as a data frame) to an output SQLite database.
}
\examples{
addDbTable("input.db", "output.db", "FVS_STANDINIT", TRUE)

}
