% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools.r
\name{addERUGroup}
\alias{addERUGroup}
\title{Add ERU Group Label (addERUGroup)}
\usage{
addERUGroup(group, eru)
}
\arguments{
\item{group}{Character string of FVS group labels.}

\item{eru}{Character string representing ERU code.}
}
\value{
Modified group string with ERU code(s) appended.
}
\description{
Adds an ERU group label to GROUPS column of FVS_STANDINIT,
FVS_PLOTINIT, FVS_STANDINIT_PLOT, FVS_STANDINIT_COND, FVS_PLOTININT_PLOT for
PPF. This function is called from dbCompile when addERU is TRUE.
}
\examples{
addERUGroup("GROUP1", "PPG")

}
