% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baStory.r
\name{baStory}
\alias{baStory}
\title{Calculate Canopy Storiedness (BA_STORY)}
\usage{
baStory(
  data,
  stand = "StandID",
  dbh = "DBH",
  expf = "TPA",
  BA = 0,
  TPA = 0,
  CC = 0,
  debug = F
)
}
\arguments{
\item{data}{Data frame containing tree records from a single stand or plot.
Must contain stand/plot ID, DBH, and expansion factor columns.}

\item{stand}{Character string for stand/plot ID column name. Default is
"StandID".}

\item{dbh}{Character string for DBH column name. Default is "DBH".}

\item{expf}{Character string for TPA column name. Default is "TPA".}

\item{BA}{Basal area per acre of plot/stand.}

\item{TPA}{Trees per acre of plot/stand.}

\item{CC}{Percent canopy cover corrected for overlap of plot/stand.}

\item{debug}{Logical; if TRUE, prints debug output to R console.}
}
\value{
Storiedness value (integer value from 0 - 3)
}
\description{
Calculates canopy layers/storiedness (BA_STORY) for a stand/plot in
accordance with the Region 3 ruleset from Vandendriesche, D., 2013. A
Compendium of NFS Regional Vegetation Classification Algorithms. USDA Forest
Service. Fort Collins, CO. (2013 pg. R3-4 - R3-5).
}
\examples{
 Example usage:
 baStory(data, stand = "StandID", dbh = "DBH", expf = "TPA", BA = 120, TPA = 300, CC = 45)
}
