% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canClass.r
\name{canClass}
\alias{canClass}
\title{Calculate Canopy Class (canClass)}
\usage{
canClass(cc, region, debug = F)
}
\arguments{
\item{cc}{Percent canopy cover corrected for overlap.}

\item{region}{USFS region code. Valid values are 1, 2, 3, 4, 5, 6, 8, 9, or 10.}

\item{debug}{Logical; if TRUE, prints debug output to R console. Default is FALSE.}
}
\value{
Integer code for canopy class (region-specific meaning).
}
\description{
Calculates canopy class based on the canopy cover percent corrected for overlap.
Canopy classes can be calculated using USFS Region 1, Region 2, Region 3, Region 4, and Region 6 rule sets described in:
Vandendriesche, D., 2013. A Compendium of NFS Regional Vegetation Classification Algorithms. USDA Forest Service. Fort Collins, CO.
}
\examples{
canClass(cc = 45, region = 3)

}
