% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canClass.r
\name{canClassR1}
\alias{canClassR1}
\title{Calculate Canopy Class for Region 1 (canClassR1)}
\usage{
canClassR1(cc, debug = F)
}
\arguments{
\item{cc}{Percent canopy cover corrected for overlap.}

\item{debug}{Logical; if TRUE, prints debug output to R console. Default is FALSE.}
}
\value{
Integer code for canopy class (0: Sparse, 1: Low, 2: Open, 3: Moderate, 4: Closed).

0: 0  - 10\%  CC (Sparse)

1: 10 - 25\%  CC (Low)

3: 40 - 60\%  CC (Moderate)

4: 60 - 100\% CC (Closed)
}
\description{
Calculates canopy class from percent canopy cover corrected for overlap using USFS R1 rule sets.
Vandendriesche, D., 2013. A Compendium of NFS Regional Vegetation Classification Algorithms. USDA Forest Service. Fort Collins, CO R1-9.
}
\examples{
canClassR1(cc = 45)

}
