% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canClass.r
\name{canClassR246}
\alias{canClassR246}
\title{Calculate Canopy Class for Regions 2, 4, 6 (canClassR246)}
\usage{
canClassR246(cc, debug = F)
}
\arguments{
\item{cc}{Percent canopy cover corrected for overlap.}

\item{debug}{Logical; if TRUE, prints debug output to R console. Default is FALSE.}
}
\value{
Integer code for canopy class (0: Sparse, 1: Low, 2: Moderate, 3: Closed).
}
\description{
Calculates canopy class from percent canopy cover corrected for overlap using USFS R2, R4, R6 rule sets.
Vandendriesche, D., 2013. A Compendium of NFS Regional Vegetation Classification Algorithms. USDA Forest Service. Fort Collins, CO R2-4, R4-6, R6-5.
}
\examples{
canClassR246(cc = 45)

}
