% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canClass.r
\name{canClassR3}
\alias{canClassR3}
\title{Calculate Canopy Class for Region 3 (canClassR3)}
\usage{
canClassR3(cc, debug = F)
}
\arguments{
\item{cc}{Percent canopy cover corrected for overlap.}

\item{debug}{Logical; if TRUE, prints debug output to R console. Default is FALSE.}
}
\value{
Integer code for canopy class (0: Sparse, 1: Low, 2: Moderate, 3: Closed).
}
\description{
Calculates canopy class from percent canopy cover corrected for overlap using USFS R3 rule sets.
Vandendriesche, D., 2013. A Compendium of NFS Regional Vegetation Classification Algorithms. USDA Forest Service. Fort Collins, CO R3-6.
}
\examples{
canClassR3(cc = 45)

}
