% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canSizCl.r
\name{canSizCl}
\alias{canSizCl}
\title{Calculate Canopy Size Class (canSizCl)}
\usage{
canSizCl(
  data,
  stand = "StandID",
  dbh = "DBH",
  expf = "TPA",
  crwidth = "CrWidth",
  TPA = 0,
  CC = 0,
  type = 1,
  debug = F
)
}
\arguments{
\item{data}{Data frame containing tree records from a single stand or plot. Must contain stand/plot ID, DBH, expansion factor, and crown width columns.}

\item{stand}{Character string for stand/plot ID column name. Default is "StandID".}

\item{dbh}{Character string for DBH column name. Default is "DBH".}

\item{expf}{Character string for expansion factor column name. Default is "TPA".}

\item{crwidth}{Character string for crown width column name. Default is "CrWidth".}

\item{TPA}{Trees per acre of stand/plot.}

\item{CC}{Percent canopy cover corrected for overlap of stand/plot.}

\item{type}{Indicator for diameter class type: 1 = Midscale mapping (default), 2 = Timberland dominance, 3 = Woodland dominance.}

\item{debug}{Logical; if TRUE, prints debug output to R console. Default is FALSE.}
}
\value{
Integer code for canopy size class.

Possible return values when type = 1

1 = seedling sapling canopy cover

2 = small tree canopy cover

3 = medium and large tree canopy cover

4 = very large tree canopy cover

5 - giant tree canopy cover

Possible return values when type = 2

1 = seedling sapling canopy cover

2 = small tree canopy cover

3 = medium tree canopy cover

4 = large - giant tree cover

Possible return values when type = 3

1 = seedling sapling canopy cover

2 = small tree canopy cover

3 = medium - giant tree cover
}
\description{
Calculates canopy size class for a stand/plot in accordance with Region 3
rulesets from Vandendriesche, D., 2013. A Compendium of NFS Regional
Vegetation Classification Algorithms. USDA Forest Service. Fort Collins, CO.
(2013 pg.R3-3 - R3-4).
}
\examples{
Example usage:
canSizCl(data, stand = "StandID", dbh = "DBH", expf = "TPA", crwidth = "CrWidth", TPA = 300, CC = 45, type = 1)

}
