% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInput.r
\name{carbonQuery}
\alias{carbonQuery}
\title{Build SQL Query for FVS Carbon Table}
\usage{
carbonQuery(cases)
}
\arguments{
\item{cases}{Character vector of case IDs.}
}
\value{
Character string of SQL statements.
}
\description{
Builds and returns a string of SQL statements to read in data from the
FVS_Carbon table. This query will read in all fields from the FVS_Carbon
table. The \code{carbonQuery} function can be invoked outside of main.R for testing
purposes and can be used in the \code{dbGetQuery} function of the RSQLite R package.
}
\examples{
carbonQuery(c("Case1", "Case2"))
}
