% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInput.r
\name{caseQuery}
\alias{caseQuery}
\title{Build SQL Query for FVS Cases Table}
\usage{
caseQuery(runTitle)
}
\arguments{
\item{runTitle}{Character string for FVS run title.}
}
\value{
Character string of SQL statements.
}
\description{
Builds and returns a string of SQL statements to read CaseID, StandID,
Groups, and Variant from the FVS_Cases table for a given run title. Can be
used with dbGetQuery from RSQLite. Added in query for Stand_CN in new verison.
}
\examples{
caseQuery("MyRunTitle")
}
