% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools.r
\name{checkDataTypes}
\alias{checkDataTypes}
\title{Check and Convert Data Frame Column Types to Match SQLite Table (checkDataTypes)}
\usage{
checkDataTypes(con, tableName, data, debug = F)
}
\arguments{
\item{con}{Connection to SQLite database.}

\item{tableName}{Name of table in SQLite database.}

\item{data}{Input data frame whose column types will be checked and converted.}

\item{debug}{Logical; if TRUE, prints debug output to R console. Default FALSE.}
}
\value{
Data frame with columns cast to correct types.
}
\description{
Checks if column data types in a data frame match the field types in a SQLite database table. Converts columns as needed.
}
\examples{
checkDataTypes(con, "FVS_STANDINIT", df)

}
