% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInput.r
\name{checkDBTables}
\alias{checkDBTables}
\title{Check FVS Output Database Tables (checkDBTables)}
\usage{
checkDBTables(con, variants, addCompute, addPotFire, addFuels, addCarbon)
}
\arguments{
\item{con}{Connection to input .db}

\item{variants}{Character vector of FVS variant codes associated with runs being processed in main function (main.R).}

\item{addCompute}{Logical. Include FVS_Compute table in output?}

\item{addPotFire}{Logical. Include FVS_Potfire table in output?}

\item{addFuels}{Logical. Include FVS_Fuels table in output?}

\item{addCarbon}{Logical. Include FVS_Carbon table in output?}
}
\value{
Character string containing message.
}
\description{
This function performs a series of checks to see if the incoming FVS output
database is ready for processing. A message is returned from the function
which will determine if processing of input database in main (main.R)
function is terminated or continued.
}
\examples{
con <- RSQLite::dbConnect(RSQLite::SQLite(), ":memory:")
checkDBTables(con, c("CS", "LS"), TRUE, TRUE, TRUE, TRUE)
}
