% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools.r
\name{collectDBPaths}
\alias{collectDBPaths}
\title{Collect SQLite Database Paths (collectDBPaths)}
\usage{
collectDBPaths(dbIn = c(), unzipDir = "")
}
\arguments{
\item{dbIn}{Character vector of directory paths and file names to SQLite databases or zipped folders.}

\item{unzipDir}{Directory path to folder where contents of zipped folders will be stored.}
}
\value{
Character vector of directory paths and file names to SQLite databases.
}
\description{
Takes a character vector of directory paths and file names to SQLite databases or zipped folders and returns an updated vector of database paths.
}
\examples{
collectDBPaths(c("db1.db", "db2.zip"))

}
