% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInput.r
\name{collectID}
\alias{collectID}
\title{Collect IDs for SQL Query}
\usage{
collectID(ids)
}
\arguments{
\item{ids}{Character vector of IDs.}
}
\value{
Character string of IDs surrounded by parentheses and separated by commas.
}
\description{
Takes in a character vector of IDs used for querying a database
and collects them into a single string where IDs are surrounded by parentheses
and separated by commas.

Example IDs: "Stand1", "Stand2", "Stand3" would be collected into the
the following string: "('Stand1', 'Stand2', 'Stand3')"
}
\examples{
collectID(c("Stand1", "Stand2", "Stand3"))
}
