% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInput.r
\name{computeQuery}
\alias{computeQuery}
\title{Build SQL Query for FVS Compute Table}
\usage{
computeQuery(cases)
}
\arguments{
\item{cases}{Character vector of case IDs.}
}
\value{
Character string of SQL statements.
}
\description{
Builds and returns a string of SQL statements to read in data from the
FVS_Compute table. Can be used with dbGetQuery from RSQLite.
}
\examples{
computeQuery(c("Case1", "Case2"))
}
