% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeVerticalStructure.r
\name{computeVerticalStructure}
\alias{computeVerticalStructure}
\title{Calculate Number of Canopy Layers (computeVerticalStructure)}
\usage{
computeVerticalStructure(BAByDiameterClass)
}
\arguments{
\item{BAByDiameterClass}{Vector containing the proportion of basal area by diameter class. Classes are 0-4.9", 5-9.9", 10-14.9", 15-19.9", 20-24.9", and 25"+.}
}
\value{
Number of Canopy layers (VERTICAL STRUCTURE)--1,2,3 or continuous (C)
}
\description{
Calculates the number of canopy layers.
}
\examples{
A single-storied stand structure:
ba_proportions_1 <- c(0.1, 0.7, 0.2, 0, 0, 0)
computeVerticalStructure(ba_proportions_1)

A two-storied stand structure:
ba_proportions_2 <- c(0.4, 0.1, 0.4, 0, 0.1, 0)
computeVerticalStructure(ba_proportions_2)

}
