% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools.r
\name{convertType}
\alias{convertType}
\title{Convert Between R and SQLite Data Types (convertType)}
\usage{
convertType(value, type = 1)
}
\arguments{
\item{value}{Character string corresponding to an R or SQLite data type.}

\item{type}{Integer; 1 = convert to SQLite type, 2 = convert to R type. Default 1.}
}
\value{
Character string corresponding to R or SQLite data type.
}
\description{
Maps R data types to SQLite data types or vice versa, depending on the type argument.
}
\examples{
convertType("double", 1)
convertType("INTEGER", 2)

}
