% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAttr.r
\name{correctCC}
\alias{correctCC}
\title{Correct Percent Canopy Cover for Overlap}
\usage{
correctCC(CC)
}
\arguments{
\item{CC}{Numeric uncorrected percent canopy cover value.}
}
\value{
Numeric corrected percent canopy cover value.
}
\description{
Converts uncorrected percent canopy cover values to corrected canopy cover
using the Crookston and Stage (1999) relationship.
}
\examples{
correctCC(35)
}
