% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools.r
\name{correctSp}
\alias{correctSp}
\title{Correct USDA Plant Symbol (correctSp)}
\usage{
correctSp(sp = "2TB")
}
\arguments{
\item{sp}{Character string corresponding to USDA plant symbol.}
}
\value{
Corrected USDA plant symbol (character).
}
\description{
Takes in a USDA plant symbol and returns a corrected symbol if input is 2TD or 2TE. Can be expanded for other corrections.
}
\examples{
correctSp("2TE")
correctSp("2TD")
correctSp("PIPO")

}
