% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customAttr.r
\name{customAttr}
\alias{customAttr}
\title{Custom Attribute Calculation (customAttr)}
\usage{
customAttr(
  data,
  region = 3,
  stand = "StandID",
  species = "SpeciesPLANTS",
  dbh = "DBH",
  ht = "Ht",
  crwidth = "CrWidth",
  expf = "TPA",
  vol1 = "MCuFt",
  vol2 = "SCuFt",
  vol3 = "SBdFt",
  vol1DBH = vol1DBH,
  vol2DBH = vol2DBH,
  vol3DBH = vol3DBH,
  customVars = customVars,
  debug = F
)
}
\arguments{
\item{data}{Data frame containing tree records from a single stand or plot. Must contain stand/plot ID, DBH, height, species, expansion factor, and crown width columns.}

\item{region}{Integer variable corresponding to USFS region number (1, 2, 3, 8 or "MPSG" are valid values).}

\item{stand}{Character string for stand/plot ID column name. Default is "StandID".}

\item{species}{Character string for USDA plant symbols column name. Default is "SpeciesPLANTS".}

\item{dbh}{Character string for DBH column name. Default is "DBH".}

\item{ht}{Character string for total tree height column name. Default is "Ht".}

\item{crwidth}{Character string for crown width column name. Default is "CrWidth".}

\item{expf}{Character string for TPA column name. Default is "TPA".}

\item{vol1}{Character string for total cubic foot (western) or merchantable cubic foot (eastern) volume column name. Default is "MCuFt".}

\item{vol2}{Character string for merchantable cubic foot (western) or sawlog cubic foot (eastern) volume column name. Default is "SCuFt".}

\item{vol3}{Character string for board foot (western) or sawlog board foot (eastern) volume column name. Default is "SBdFt".}

\item{debug}{Logical; if TRUE, prints debug output to R console. Default is FALSE.}

\item{custVars}{Dataframe specifying the arguments for the custom variables. Arguments include:
\itemize{
\item VARIABLE_NAME: name of the custom variable to be returned (required)
\item ATTRIBUTE: variable to be returned (required)
\item SPECIES: 2-letter species code or custom species group name (default=ALL)
\item MIN_DBH: Smallest DBH in inches to be included (default=0.0)
\item MAX_DBH: Largest DBH in inches to be included (default=200.0)
\item MIN_HT:  Shortest tree in feet to be included (default=0.0)
\item MAX_HT:  Tallest tree in feet to be included (default=500.0)
}}
}
\value{
Named list containing custom attributes for each plot/stand.
}
\description{
Provides similar functionality to the SPMCDBH() FVS Event Monitor function by returning the following attributes by species or species group, and ranges of diameters and heights:
\itemize{
\item Trees per acre
\item Basal area per acre
\item Reineke SDI
\item Zeide SDI
\item Percent canopy cover
\item Quadratic mean diameter
\item Basal area weighted diameter
\item Average height
\item Total cubic foot volume/acre
\item Merch cubic foot volume/acre
\item Board foot volume/acre
}
}
\examples{
customAttr(data, region = 3, stand = "StandID", species = "SpeciesPLANTS", dbh = "DBH", ht = "Ht", crwidth = "CrWidth", expf = "TPA")

}
