% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools.r
\name{dbCompile}
\alias{dbCompile}
\title{Compile and Merge FVS SQLite Databases (dbCompile)}
\usage{
dbCompile(
  dbIn = NULL,
  dbOut = NULL,
  dbTables = c("FVS_STANDINIT", "FVS_TREEINIT", "FVS_PLOTINIT", "FVS_STANDINIT_PLOT",
    "FVS_STANDINIT_COND", "FVS_PLOTINIT_PLOT", "FVS_TREEINIT_PLOT", "FVS_TREEINIT_COND"),
  buildGaak = T,
  gaakType = 2,
  addERU = T,
  deleteInput = F,
  readChunks = F,
  rowsToRead = 5000
)
}
\arguments{
\item{dbIn}{Character vector of directory paths and file names for SQLite databases to process. Files can be .db or .zip containing .db files.}

\item{dbOut}{Character string for SQLite database to write out to.}

\item{dbTables}{Character vector of database tables to process from dbIn. Default includes all standard FVS tables.}

\item{buildGaak}{Logical; if TRUE, writes FVS_GROUPADDFILESANDKEYWORDS to dbOut. Default TRUE.}

\item{gaakType}{Integer (1-3) for GAAK table type. See fvsGaak for details. Default 2.}

\item{addERU}{Logical; if TRUE, adds ERU as a field and grouping code. Default TRUE.}

\item{deleteInput}{Logical; if TRUE, deletes input databases after processing. Default FALSE.}

\item{readChunks}{Logical; if TRUE, reads data in chunks to reduce RAM usage. Default FALSE.}

\item{rowsToRead}{Integer; number of rows to read per chunk if readChunks is TRUE. Default 5000.}
}
\value{
Message indicating that database has been created.
}
\description{
Reads in database tables from input FVS-ready data sets and writes the
database tables from each to a single output SQLite database. Can combine
input FVS databases or extract FVS database tables from a larger database
such as those on the FIA datamart.
}
\examples{
dbCompile(dbIn = "input.db", dbOut = "output.db", dbTables = c("FVS_STANDINIT", "FVS_TREEINIT","FVS_PLOTINIT", buildGaak = T)

}
