% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools.r
\name{deleteFiles}
\alias{deleteFiles}
\title{Delete Files or Directories (deleteFiles)}
\usage{
deleteFiles(files = c(), recur = FALSE)
}
\arguments{
\item{files}{Character vector of directory paths and file names to delete.}

\item{recur}{Logical; if TRUE, deletes recursively. Default FALSE.}
}
\value{
Integer value of 0 (invisible).
}
\description{
Deletes files or directories specified in a character vector.
}
\examples{
deleteFiles(c("file1.db", "file2.db"), recur = TRUE)

}
