% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denSize.r
\name{denSizeR8}
\alias{denSizeR8}
\title{Calculate Density/Size Class for USFS R8}
\usage{
denSizeR8(data, stand = "StandID", attrList, debug = F)
}
\arguments{
\item{data}{Data frame containing tree records for a single stand or plot. Must include a stand/plot ID column.}

\item{stand}{Character string for the name of the stand/plot ID column in data. Default is "StandID".}

\item{attrList}{List of size and density attributes for species in the stand/plot, as produced by plotAttr().}

\item{debug}{Logical; if TRUE, prints debug output to the R console.}
}
\value{
Character string with the density/size class code, or NA if not applicable.
}
\description{
Calculates a density/size class based on USFS R8 rulesets defined
by Chad Keyser in 2023. Size density class criteria for R8 is as follows:

If the basal area of the stand is less than 10 BA, then the size class in the
size density classification is advanced regeneration.

Advanced regeneration size class = 1

If the size class of the stand is advanced regeneration, then one of the
following density classes will be determined:
\itemize{
\item A: 0 - 200 stand TPA
\item B: 200 - 400 stand TPA
\item C: 400 - 600 stand TPA
\item D: 600 + stand TPA
}

If the basal area of the stand is greater than or equal to 10 BA, then the size
class in the size density classification is either non-merch size, pulpwood
size, or saw timber size. The size class with the most BA is selected as the
size class in the density/size classification.
\itemize{
\item Non-merch size class = 2
\item Pulpwood size class = 3
\item Saw timber size class = 4
}

If the size class of the stand is non-merch, pulpwood or saw timber, then one
of the following density classes will be determined:
\itemize{
\item A: 0 - 40 BA
\item B: 40 - 80 BA
\item C: 80 - 120 BA
\item D: 120+ BA
}
}
\examples{
attrList <- list(ALL = c(TPA=300, BA=50, NMBA=10, PWBA=20, STBA=20, ARTPA=0))
df <- data.frame(StandID="A", X=1)
denSizeR8(df, attrList=attrList)
}
