% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domType.r
\name{domTypeR3}
\alias{domTypeR3}
\title{Calculate Stand Dominance Type (Region 3)}
\usage{
domTypeR3(
  data,
  stand = "StandID",
  species = "SpeciesPLANTS",
  dbh = "DBH",
  expf = "TPA",
  crwidth = "CrWidth",
  TPA = 0,
  CC = 0,
  debug = F
)
}
\arguments{
\item{data}{Data frame containing tree records from a single stand or plot. Data frame must contain a column corresponding to stand/plot ID, DBH, species code (USDA plant symbol), expansion factor, and crown width for each tree record.}

\item{stand}{Character string corresponding to name of column pertaining to stand or plot ID associated with tree records in data argument. By default, this value is set to "StandID".}

\item{species}{Character string corresponding to name of column pertaining to USDA plant symbols of tree records in data argument. By default, this argument is set to "SpeciesPLANTS".}

\item{dbh}{Character string corresponding to name of column pertaining to DBH of tree records in data argument. By default, this argument is set to"DBH".}

\item{expf}{Character string corresponding to name of column pertaining to TPA of tree records in data argument. By default, this argument is set to "TPA".}

\item{crwidth}{Character string corresponding to name of column pertaining to crown width values of tree records in data argument. By default, this argument is set to "CrWidth".}

\item{TPA}{Numeric. Trees per acre for the stand/plot.}

\item{CC}{Numeric. Canopy cover (uncorrected for overlap).}

\item{debug}{Logical; if TRUE, prints debug output.}
}
\value{
Named list containing:
\item{DOMTYPE}{Dominance type}
\item{DCC1}{Dominant species/genus/category or species/genus/category occurring before underscore in dominance type}
\item{XDCC1}{Percent canopy cover corrected for overlap represented by DCC1}
\item{DCC2}{Species/genus/category occurring after underscore in dominance type}
\item{XDCC2}{Percent canopy cover corrected for overlap represented by DCC2}
}
\description{
Calculates stand dominance type using USFS Region 3 rule sets in accordance
with Vandendriesche, D., 2013. A Compendium of NFS Regional Vegetation
Classification Algorithms. USDA Forest Service. Fort Collins, CO (R3-1 -
R3-3).
}
\examples{
domTypeR3(df, TPA=100, CC=30)
}
