% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domType.r
\name{domTypeR8}
\alias{domTypeR8}
\title{Calculate Stand Dominance Type (Region 8)}
\usage{
domTypeR8(data, stand = "StandID", attrList, debug = F)
}
\arguments{
\item{data}{Data frame of tree records for a single stand or plot. Must include a stand/plot ID column.}

\item{stand}{Name of the stand/plot ID column. Default is "StandID".}

\item{attrList}{List of size and density attributes for species, as produced by \code{plotAttr()}.}

\item{debug}{Logical; if TRUE, prints debug output.}
}
\value{
Named list containing:
\item{SSDOMSPP}{Dominance type of advanced regeneration}
\item{NMDOMSPP}{Dominance type of non-merchantable trees}
\item{PWDOMSPP}{Dominance type of pulpwood sized trees}
\item{STDOMSPP}{Dominance type of saw timber sized trees}
\item{DOMTYPE}{Dominance type of stand}
}
\description{
Calculates stand dominance type using USFS Region 8 rule sets determined by
Chad Keyser in 2023. This R8 classification is not the official rule set for
USFS R8. It was developed for a specific project in North Carolina. Five dominance
type classifications are calculated:
advanced regeneration, non-merchantable trees, pulpwood sized trees, saw
timber sized trees, and the entire stand. Criteria for
determining dominance type is as follows:

Advance regeneration dominance
\itemize{
\item If 1 species represents greater than 70\% of trees per acre in advanced
regeneration size class then dominance type is a single species.
\item If 2 species represent greater than 70\% of trees per acre in advanced
regeneration size class then dominance type is two species.
\item Otherwise dominance type is 3 species (3 most abundant species based on TPA).
\item If there is no TPA in advance regeneration size class, then dominance type is
set to NONE.
}

Non merch, pulpwood, sawtimber, and stand dominance
\itemize{
\item If 1 species represents greater than 70\% of basal area in respective size class
(non merch, pulpwood, sawtimber, and stand) the dominance type is two species.
\item If 2 species represent greater than 70\% of basal area in respective size class
(non merch, pulpwood, sawtimber, and stand) the dominance type is two species.
\item Otherwise dominance type is 3 species (3 most abundant species based on BA).
\item If there is no BA in a given size class, then dominance type is set to NONE
for that size class.
}
}
\examples{
attrList <- list(ALL = c(SSTPA=100, NMBA=50, PWBA=30, STBA=20, BA=200, TPA=300),
               PIPO = c(SSTPA=60, NMBA=30, PWBA=10, STBA=10, BA=100),
               ABGR = c(SSTPA=40, NMBA=20, PWBA=20, STBA=10, BA=100))
df <- data.frame(StandID="A", X=1)
domTypeR8(df, attrList=attrList)
}
