% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools.r
\name{eru}
\alias{eru}
\title{Get USFS Region 3 ERU Codes (eru)}
\usage{
eru()
}
\value{
Character vector containing USFS R3 ERU Codes.
}
\description{
Returns a vector of USFS Region 3 ERU codes. This function is
called in pvConvert to crosswalk PV Codes to ERU.
}
\examples{
eru()

}
