% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domType.r
\name{excGenusSp}
\alias{excGenusSp}
\title{Find Mutually Exclusive Genus Index}
\usage{
excGenusSp(sp, genNames, genusList)
}
\arguments{
\item{sp}{Name of the most abundant species by percent canopy cover for current stand in \code{domType()}}

\item{genusList}{Vector of genera whose indices are named by each species found in current stand in \code{domType()}}

\item{spNames}{Vector of genera names ordered by abundance in terms of percent canopy cover for current stand in \code{domType()}}
}
\value{
Index where the name of the most abundant genus does not match the genus in
the input list of genera.
}
\description{
Used in the Region 3 dominance type algorithm (see domTypeR3) to determine a genus that is mutually exclusive from
the genus of the most abundant species by percent canopy cover in lead 14 of
R3 dominance type algorithm. This function takes the name of the most
abundant species, a vector with names of each genus (sorted by genera with most
to least canopy cover), and a named vector of genera (not sorted by percent canopy cover).
}
\examples{
excGenusSp("PIPO", c("Pinus", "Abies"), c(PIPO="Pinus", ABGR="Abies"))
}
