% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInput.r
\name{fuelsQuery}
\alias{fuelsQuery}
\title{Build SQL Query for FVS Fuels Table}
\usage{
fuelsQuery(cases)
}
\arguments{
\item{cases}{Character vector of case IDs.}
}
\value{
Character string of SQL statements.
}
\description{
Builds and returns a string of SQL statements to read in data from the
FVS_Fuels table. This query will read in all fields from the FVS_Fuels
table. The \code{fuelsQuery} function can be invoked outside of main.R for testing
purposes and can be used in the \code{dbGetQuery} function of the RSQLite R package.
}
\examples{
fuelsQuery(c("Case1", "Case2"))
}
