% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools.r
\name{fvsDBTable}
\alias{fvsDBTable}
\title{Check if Table Name is FVS Input Table (fvsDBTable)}
\usage{
fvsDBTable(dbTable = "")
}
\arguments{
\item{dbTable}{Character string for database table name.}
}
\value{
TRUE if database table name is an input FVS table name, FALSE otherwise.
}
\description{
Checks if a database table name is an input FVS table name.
\itemize{
\item FVS tables:
\item "FVS_STANDINIT"
\item "FVS_TREEINIT"
\item "FVS_PLOTINIT"
\item "FVS_STANDINIT_PLOT"
\item "FVS_STANDINIT_COND"
\item "FVS_PLOTINIT_PLOT"
\item "FVS_TREEINIT_PLOT"
\item "FVS_TREEINIT_COND"
}
}
\examples{
fvsDBTable("FVS_STANDINIT")

}
