% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools.r
\name{fvsGaak}
\alias{fvsGaak}
\title{Create FVS GroupAddFilesAndKeywords Table (fvsGaak)}
\usage{
fvsGaak(dbName = "FVS_Data", type = 2)
}
\arguments{
\item{dbName}{Character string corresponding to name of database used in GAAK table. Default is "FVS_Data". A .db extension will be appended.}

\item{type}{Integer to determine what grouping codes are included in GAAK table. 1 = Standard FVS grouping codes, 2 = FIA grouping codes, 3 = Both. Default is 2.}
}
\value{
Data frame containing the GAAK table.
}
\description{
Creates a FVS_GroupAddFilesAndKeyword (GAAK) table with the appropriate SQL statements for a variety of grouping codes.
}
\examples{
fvsGaak(dbName = "FVS_Data", type = 2)

}
