% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools.r
\name{fvsGetCols}
\alias{fvsGetCols}
\title{Get FVS Input Database Column Names (fvsGetCols)}
\usage{
fvsGetCols()
}
\value{
Character vector of FVS column names.
}
\description{
Returns a character vector of all column names that can be included in a FVS input database.
}
\examples{
fvsGetCols()

}
