% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools.r
\name{fvsGetTypes}
\alias{fvsGetTypes}
\title{Get FVS Input Database Column Data Types (fvsGetTypes)}
\usage{
fvsGetTypes()
}
\value{
Character vector of FVS field data types.
}
\description{
Returns a character vector of data types associated with values returned from fvsGetCols().
}
\examples{
fvsGetTypes()

}
