% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canSizCl.r
\name{getCanSizeDC}
\alias{getCanSizeDC}
\title{Get Canopy Size Diameter Class (getCanSizeDC)}
\usage{
getCanSizeDC(DBH, type = 1, debug = F)
}
\arguments{
\item{DBH}{DBH value.}

\item{type}{Indicator for diameter class type: 1 = Midscale mapping
(default), 2 = Timberland dominance, 3 = Woodland dominance.}

\item{debug}{Logical; if TRUE, prints debug output to R console. Default is
FALSE.}
}
\value{
Integer code for diameter class (see documentation for possible
values by type).

Possible return values when type = 1

1 = seedling sapling canopy cover

2 = small tree canopy cover

3 = medium and large tree canopy cover

4 = very large tree canopy cover

5 - giant tree canopy cover

Possible return values when type = 2

1 = seedling sapling canopy cover

2 = small tree canopy cover

3 = medium tree canopy cover

4 = large - giant tree cover

Possible return values when type = 3

1 = seedling sapling canopy cover

2 = small tree canopy cover

3 = medium - giant tree cover
}
\description{
Returns a diameter class used to calculate canopy size class in the canSizCl
function.
}
\examples{
getCanSizeDC(DBH = 12, type = 1)

}
