% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools.r
\name{getDataTypes}
\alias{getDataTypes}
\title{Get Data Types for SQLite Table Fields (getDataTypes)}
\usage{
getDataTypes(con, dbTable = "")
}
\arguments{
\item{con}{Connection to SQLite database.}

\item{dbTable}{Character string for database table name.}
}
\value{
Named vector of field names and data types.
}
\description{
Returns a named vector of field names and associated data types for all fields in a SQLite database table.
}
\examples{
getDataTypes(con, "FVS_STANDINIT")

}
