% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInput.r
\name{getGroup}
\alias{getGroup}
\title{Extract Group from FVS Group Labels}
\usage{
getGroup(groups, label)
}
\arguments{
\item{groups}{String containing group labels as formatted in the FVS_Cases table.}

\item{label}{Label used to extract the group of interest (e.g., "ERU=").}
}
\value{
Extracted group string or "ALL" if not found.
}
\description{
Not in use.
Takes a string of FVS group labels and returns the group following the input string identifier as specified in the label argument.
}
\examples{
getGroup("ERU=MCW,AGE=30", "ERU=")
}
