% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInput.r
\name{getIndexNames}
\alias{getIndexNames}
\title{Get Index Names from SQLite Database}
\usage{
getIndexNames(input)
}
\arguments{
\item{input}{Connection to a SQLite database (.db)}
}
\value{
Character vector of index names that exist in input argument.
}
\description{
Returns the names of indices that exist in the input SQLite database.
}
\examples{
con <- RSQLite::dbConnect(RSQLite::SQLite(), ":memory:")
getIndexNames(con)
}
