% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInput.r
\name{getPotFireVars}
\alias{getPotFireVars}
\title{Get FVS_POTFIRE Table Variable Names}
\usage{
getPotFireVars()
}
\value{
Character vector of FVS_POTFIRE variable names.
}
\description{
Returns the variable names included in the FVS_POTFIRE database table except
for StandID. Used to order columns for combining with FVS_POTFIRE_EAST output
so it can be combined with output from the standard FVS_POTFIRE_TABLE.
}
\examples{
getPotFireVars()
}
