% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSS.r
\name{HSS}
\alias{HSS}
\title{Calculate Habitat Structural Stage (HSS)}
\usage{
HSS(HSStype, data, TPA, CC, plotvals)
}
\arguments{
\item{HSStype}{Integer. Habitat structural stage type (1 = 1-4C type, 2 = 1-5 type).}

\item{data}{Data frame containing tree records from a single stand or plot. Must include columns for stand/plot ID, DBH, species code (USDA plant symbol), expansion factor, and height.}

\item{TPA}{Numeric. Trees per acre for the stand/plot.}

\item{CC}{Numeric. Percent canopy cover of the stand/plot.}

\item{plotvals}{Named list containing plot attributes for each species and/or across all species in the plot/stand.}
}
\value{
Named list containing:
\item{DOM_TYPE_R2}{Dominance type reported as the top 3 species}
\item{DOM_TYPE_R2_CC1}{Percent canopy cover of the most dominant species}
\item{DOM_TYPE_R2_CC2}{Percent canopy cover of the 2nd most dominant species}
\item{DOM_TYPE_R2_CC3}{Percent canopy cover of the 3rd most dominant species}
\item{COVERTYPE_R2}{R2 cover type}
\item{TREE_SIZE_CLASS_R2}{R2 tree size class}
\item{CROWN_CLASS_R2}{R2 percent canopy cover class}
\item{HSS1_4C}{Habitat structural stage for 1 through 4C classes}
\item{HSS1_5}{Habitat structural stage for 1 through 5 classes}
}
\description{
Calculates two different habitat structural stage variables for a stand or
plot. One is based on the logic found in the R2Calcs (R2.r) and the other
is based on the hss_wi5.kcp algorithm (see
https://www.fs.usda.gov/fvs/software/addfiles/hss_wi5.zip). The function
returns a named list with dominance type, canopy cover, cover type, tree size
class, crown class, and habitat structural stage for both the 1-4C and 1-5
class systems.
}
\examples{
 HSS(1, data, TPA=100, CC=30, plotvals=plotvals)
}
