% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPSG.r
\name{MPSG}
\alias{MPSG}
\title{Calculate MPSG Standardized Vegetation Variables}
\usage{
MPSG(
  data,
  stand = "StandID",
  species = "SpeciesPLANTS",
  dbh = "DBH",
  expf = "TPA",
  ht = "Ht",
  crwidth = crwidth,
  TPA,
  BA,
  CC,
  plotvals,
  InvDB,
  InvStandTbl,
  MPSGcovTyp,
  addHSS,
  debug = F
)
}
\arguments{
\item{data}{Data frame containing tree records from a single stand or plot. Must include columns for stand/plot ID, DBH, species code (USDA plant symbol), expansion factor, and height.}

\item{stand}{Name of the column for stand or plot ID in data. Default is "StandID".}

\item{species}{Name of the column for USDA plant symbols in data. Default is "SpeciesPLANTS".}

\item{dbh}{Name of the column for DBH in data. Default is "DBH".}

\item{expf}{Name of the column for TPA (expansion factor) in data. Default is "TPA".}

\item{ht}{Name of the column for tree height in data. Default is "Ht".}

\item{crwidth}{Name of the column for crown width in data. Default is "CrWidth".}

\item{TPA}{Trees per acre for the stand/plot.}

\item{BA}{Basal area for the stand/plot.}

\item{CC}{Percent canopy cover for the stand/plot.}

\item{plotvals}{Named list of plot attributes for each species and/or all species in the plot/stand.}

\item{InvDB}{Connection to inventory database.}

\item{InvStandTbl}{Name of the stand data table in InvDB.}

\item{MPSGcovTyp}{Integer for USFS region number whose algorithms are used for COVERTYPE_MPSG calculation.}

\item{addHSS}{Logical; if TRUE, includes USFS Region 2 Habitat Structural Stage (HSS) variables in output.}

\item{debug}{Logical; if TRUE, prints debug output to R console.}
}
\value{
A named list containing:
\item{COVERTYPE_MPSG}{MPSG cover type}
\item{TREE_SIZE_CLASS}{MPSG tree size class}
\item{CROWN_CLASS}{MPSG percent canopy cover class}
\item{VERTICAL_STRUCTURE}{Number of canopy layers}
\item{HSS1_4C}{Habitat structural stage (1 through 4C classes), if addHSS = TRUE}
\item{HSS1_5}{Habitat structural stage (1 through 5 classes), if addHSS = TRUE}
}
\description{
Calculates standardized vegetation classification and structural variables for use by the Mountain Planning Services Group (MPSG).
}
\examples{
df <- data.frame(
  StandID = rep("S1", 5),
  SpeciesPLANTS = c("PIPO", "PIPO", "POTR5", "ABLA", "PICO"),
  DBH = c(12, 15, 8, 10, 20),
  TPA = c(50, 60, 40, 30, 20),
  Ht = c(30, 35, 25, 28, 40),
  CrWidth = c(10, 12, 8, 9, 15)
)
plotvals <- list()
MPSG(
  data = df,
  stand = "StandID",
  species = "SpeciesPLANTS",
  dbh = "DBH",
  expf = "TPA",
  ht = "Ht",
  crwidth = "CrWidth",
  TPA = 200,
  BA = 120,
  CC = 45,
  plotvals = plotvals,
  InvDB = NULL,
  InvStandTbl = NULL,
  MPSGcovTyp = 2,
  addHSS = FALSE,
  debug = TRUE
)
}
