% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAttr.r
\name{plotAttr}
\alias{plotAttr}
\title{Calculate Plot and Species-Level Attributes}
\usage{
plotAttr(
  data,
  region = 3,
  stand = "StandID",
  species = "SpeciesPLANTS",
  dbh = "DBH",
  ht = "Ht",
  crwidth = "CrWidth",
  expf = "TPA",
  vol1 = "MCuFt",
  vol2 = "SCuFt",
  vol3 = "SBdFt",
  min = 0,
  max = 999,
  debug = F
)
}
\arguments{
\item{data}{Data frame containing tree records from a single stand or plot.
Must contain stand/plot ID, DBH, height, species, expansion factor, and
crown width columns.}

\item{region}{Integer USFS region number (1, 2, 3, 8, or MPSG (Mountain Planning Services Group, R1-4)).}

\item{stand}{Character name of stand/plot ID column. Default is "StandID".}

\item{species}{Character name of species column (USDA plant symbol).
Default is "SpeciesPLANTS".}

\item{dbh}{Character name of DBH column. Default is "DBH".}

\item{ht}{Character name of total height column. Default is "Ht".}

\item{crwidth}{Character name of crown width column. Default is "CrWidth".}

\item{expf}{Character name of expansion factor (TPA) column. Default is "TPA".}

\item{vol1}{Character name of column pertaining to total cubic foot (wetern variants)
or merchantable cubic foot volume (eastern variants) of tree records. Default is "MCuFt".
Currently vol1 argument is only used if region argument is set to 8.}

\item{vol2}{Character name of column pertaining to merchantable cubic foot (western variants) or sawlog cubic foot
volume (eastern variants) of tree records in data argument. Default is "SCuFt". Currently vol2 is only
used if region argument is set to 8.}

\item{vol3}{Character name of column pertaining to
board foot (western variants) or sawlog board foot volume (eastern
variants) of tree records in data argument. Default is "SBdFt". Currently vol3 is only used if region
argument is set to 8.}

\item{min}{Minimum DBH to include in calculations. Default is 0.}

\item{max}{Maximum DBH to include in calculations. Default is 999.}

\item{debug}{Logical; if \code{TRUE}, prints debug output. Default is \code{FALSE}.}
}
\value{
Named list containing one element per species plus \code{ALL}. Each list element is a named numeric vector with computed plot attributes:
\item{N}{Number of records in plot/stand}
\item{BA}{Basal Area}
\item{TPA}{Trees per acre}
\item{QMD}{Quadratic mean diameter}
\item{UNCC}{Percent canopy cover uncorrected for overlap}
\item{CC}{Percent canopy cover corrected for overlap}
\item{ZSDI}{Zeide Stand Density Index}
\item{RSDI}{Reineke Stand Density Index}
\item{BA_WT_DIA}{Basal area weighted diameter}
\item{BA_WT_HT}{Basal area weighted height}

For Region 1, average height (\code{AVE_HT}) is also calculated.

For Regions 8 and 9, additional size-class attributes are calculated:
\item{SSSIZE}{TPA weighted average height of advanced regeneration size class}
\item{SSTPA}{Trees per acre of advanced regeneration size class}
\item{SSBA}{Basal area of advanced regeneration size class}
\item{NMSIZE}{Basal area weighted diameter of non merchantable size class}
\item{NMBA}{Basal area of non merchantable size class}
\item{PWSIZE}{Basal area weighted diameter of pulpwood size class}
\item{PWBA}{Basal area of pulpwood size class}
\item{STSIZE}{Basal area weighted diameter of saw timber size class}
\item{STBA}{Basal area of saw timber size class}
}
\description{
Calculates plot/stand attributes for each species and for all species
combined. Depending on region, additional attributes are calculated. See script (plotAttr.R) for more detailed documentation.
}
\examples{
df <- data.frame(
  StandID = c("S1", "S1"),
  SpeciesPLANTS = c("PIPO", "ABGR"),
  DBH = c(10, 14),
  Ht = c(45, 55),
  CrWidth = c(12, 14),
  TPA = c(60, 40),
  MCuFt = c(8, 12),
  SCuFt = c(3, 7),
  SBdFt = c(20, 40)
)
plotAttr(df, region = 3)
}
