% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInput.r
\name{potFireQuery}
\alias{potFireQuery}
\title{Build SQL Query for FVS PotFire Table}
\usage{
potFireQuery(cases, variant)
}
\arguments{
\item{cases}{Character vector of case IDs.}

\item{variant}{Two-character variant code This code is used to determine what potential fire database table data should be read from (FVS_Potfire or FVS_Potfire_East). Valid variant codes: "AK", "BM", "CA", "CI", "CR", "CS", "EC", "EM",  "IE", "KT", "LS", "NC", "NE", "PN", "SN", "SO", "TT", "UT", "WC", "WS", "OC", "OP"}
}
\value{
Character string of SQL statements.
}
\description{
Builds and returns a string of SQL statements to read in data from the
FVS_Potfire or FVS_PotFire_East table depending on the variant specified. Can
be used with dbGetQuery from RSQLite.

This query will read in the following fields from the FVS_PotFire table:

"CaseID, "Year", "Surf_Flame_Sev", "Surf_Flame_Mod", "Tot_Flame_Sev",
"Tot_Flame_Mod", "Fire_Type_Sev", "Fire_Type_Mod", "PTorch_Sev", "PTorch_Mod",
"Torch_Index", "Crown_Index", "Canopy_Ht", "Canopy_Density", "Mortality_BA_Sev",
"Mortality_BA_Mod", "Mortality_VOL_Sev", "Mortality_VOL_Mod", "Pot_Smoke_Sev",
"Pot_Smoke_Mod"

This query will read in the following fields from the FVS_PotFire_East table:

"CaseID, "Year", "Flame_Len_Sev", "Flame_Len_Mod", "Canopy_Ht",
"Canopy_Density", "Mortality_BA_Sev", "Mortality_BA_Mod", "Mortality_VOL_Sev",
"Mortality_VOL_Mod", "Pot_Smoke_Sev", "Pot_Smoke_Mod"
}
\examples{
potFireQuery(c("Case1", "Case2"), "CS")
}
