% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools.r
\name{pvCodes}
\alias{pvCodes}
\title{Get USFS Region 3 Plant Association Codes (pvCodes)}
\usage{
pvCodes()
}
\value{
Character vector containing PV Codes.
}
\description{
Returns a vector of USFS Region 3 Plant Association codes (PV codes).
This function is called in pvConvert to crosswalk PV Code to ERU.
}
\examples{
pvCodes()

}
