% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools.r
\name{pvConvert}
\alias{pvConvert}
\title{Convert PV Code to ERU Code (pvConvert)}
\usage{
pvConvert(pv)
}
\arguments{
\item{pv}{Input PV code.}
}
\value{
ERU code (character) or NA if not found.
}
\description{
Converts input Plant Association code (PV code) to USFS Region 3 ERU code. If the input PV Code is not recognized, returns NA.
}
\examples{
pvConvert("012330")

}
