% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmdSizCl.r
\name{qmdSizCl}
\alias{qmdSizCl}
\title{Classify QMD Size Class}
\usage{
qmdSizCl(qmd, debug = F)
}
\arguments{
\item{qmd}{Numeric quadratic mean diameter for the stand/plot.}

\item{debug}{Logical; if \code{TRUE}, prints debug output. Default is \code{FALSE}.}
}
\value{
Integer QMD size class code:
\item{0}{0 to <1 inch (Nonstocked)}
\item{1}{1 to <5 inches (Seed/Sap)}
\item{2}{5 to <10 inches (Small Tree)}
\item{3}{10 to <15 inches (Medium Tree)}
\item{4}{15 to <20 inches (Large Tree)}
\item{5}{20+ inches (Very Large Tree)}
}
\description{
Calculates quadratic mean diameter (QMD) size class for stand/plot using QMD for seedlings
and saplings. This value is calculated using USGS region 6 rule sets described
in Vandendriesche, D., 2013. A Compendium of NFS Regional Vegetation
Classification Algorithms. USDA Forest Service. Fort Collins, CO
}
\examples{
qmdSizCl(0.7)
qmdSizCl(12.5)
}
