% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInput.r
\name{removeCaseIndices}
\alias{removeCaseIndices}
\title{Remove CaseID Indices from FVS Tables}
\usage{
removeCaseIndices(input)
}
\arguments{
\item{input}{Connection to a SQLite database (.db)}
}
\value{
Invisibly returns 0 if successful, 1 if no tables found.
}
\description{
Removes all indices created by \code{setCaseIndices()} in FVS
database tables found in input database argument. Some Logic in this function
was borrowed from \code{mkDBIndices()} (fvsRunUtilities.R).
}
\examples{
con <- RSQLite::dbConnect(RSQLite::SQLite(), ":memory:")
removeCaseIndices(con)
}
