% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInput.r
\name{setCaseIndices}
\alias{setCaseIndices}
\title{Set CaseID Indices in FVS Tables}
\usage{
setCaseIndices(input)
}
\arguments{
\item{input}{Connection to a SQLite database (.db)}
}
\value{
Invisibly returns 0 if successful, 1 if no tables found.
}
\description{
Creates indices on the CaseID field for all FVS database tables found in the
input database. Should be used if input database in \code{main()} (main.R) was not
created in the FVS interface (i.e rFVS or FVS run through command line). Some
Logic in this function was borrowed from mkDBIndices function
(fvsRunUtilities.R). This had been set to automatically run
}
\examples{
 con <- RSQLite::dbConnect(RSQLite::SQLite(), ":memory:")
 setCaseIndices(con)
}
