% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools.r
\name{setDataTypes}
\alias{setDataTypes}
\title{Set Data Types for Data Frame Columns (setDataTypes)}
\usage{
setDataTypes(data, cols = NULL, colTypes = NULL, debug = F)
}
\arguments{
\item{data}{Input data frame.}

\item{cols}{Character vector of variable names. Default NULL uses fvsGetCols().}

\item{colTypes}{Character vector of data types for variables. Default NULL uses fvsGetTypes().}

\item{debug}{Logical; if TRUE, prints debug output to R console. Default FALSE.}
}
\value{
Data frame with columns cast to correct types.
}
\description{
Not currently in use. Checks and sets data types for columns in a data frame to match specified types. Only double, integer, and character values are considered.
}
\examples{
setDataTypes(data.frame(A=1,B="2"), cols=c("A","B"), colTypes=c("integer","character"))

}
