% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInput.r
\name{treeQuery}
\alias{treeQuery}
\title{Build SQL Query for FVS TreeList Table (treeQuery)}
\usage{
treeQuery(cases, variant)
}
\arguments{
\item{cases}{Character vector of case IDs.}

\item{variant}{Two-character variant code. This code is used to determine which FVS_TreeList columns are selected. Valid variant codes: "AK", "BM", "CA", "CI", "CR", "CS", "EC", "EM", "IE", "KT", "LS", "NC", "NE", "PN", "SN", "SO", "TT", "UT", "WC", "WS", "OC", "OP".}
}
\value{
Character string of SQL statements.
}
\description{
Builds and returns a string of SQL statements to read data from the FVS_TreeList table, with selected columns depending on the variant specified. Can be used with dbGetQuery from RSQLite.

The query will read in the following fields from the FVS_TreeList table when
variant argument is not CS, LS, NE, or SN:

"CaseID", "StandID", "Year", "SpeciesPLANTS", "MortPA", "TPA", "DBH", "Ht",
"CrWidth", "TCuFt", "MCuFt", "BdFt"

The query will read in the following fields from the FVS_TreeList table
when variant argument is CS, LS, NE, or SN:

"CaseID", "StandID", "Year", "SpeciesPLANTS", "MortPA", "TPA", "DBH", "Ht",
"CrWidth", "MCuFt", "SCuFt", "SBdFt"
}
\examples{
treeQuery(c("Case1", "Case2"), "CS")
}
